github-automation/requirements.txt
PyGithub>=2.1.1
python-dotenv>=0.19.0
pyyaml>=6.0.1
requests>=2.26.0